// codegen/crm/timeline/model/batchInputTimelineEvent.ts
/**
 * Timeline events
 * This feature allows an app to create and configure custom events that can show up in the timelines of certain CRM object like contacts, companies, tickets, or deals. You\'ll find multiple use cases for this API in the sections below.
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { TimelineEvent } from './timelineEvent';

/**
* Used to create timeline events in batches.
*/
export class BatchInputTimelineEvent {
    /**
    * A collection of timeline events we want to create.
    */
    'inputs': Array<TimelineEvent>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<TimelineEvent>"
        }    ];

    static getAttributeTypeMap() {
        return BatchInputTimelineEvent.attributeTypeMap;
    }
}