// codegen/crm/timeline/model/collectionResponseTimelineEventTemplate.ts
/**
 * Timeline events
 * This feature allows an app to create and configure custom events that can show up in the timelines of certain CRM objects like contacts, companies, tickets, or deals. You\'ll find multiple use cases for this API in the sections below.
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { Paging } from './paging';
import { TimelineEventTemplate } from './timelineEventTemplate';

/**
* Container for all templates being returned.
*/
export class CollectionResponseTimelineEventTemplate {
    /**
    * A collection of templates.
    */
    'results': Array<TimelineEventTemplate>;
    'paging'?: Paging;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "results",
            "baseName": "results",
            "type": "Array<TimelineEventTemplate>"
        },
        {
            "name": "paging",
            "baseName": "paging",
            "type": "Paging"
        }    ];

    static getAttributeTypeMap() {
        return CollectionResponseTimelineEventTemplate.attributeTypeMap;
    }
}